+++ src/components/form/form.tsx
@@ -1,4 +1,4 @@
-import { Component, Event, EventEmitter, Method, h } from '@stencil/core';
+import { Component, Event, EventEmitter, Method, Prop, h } from '@stencil/core';
 
 interface FormControl {
   tag: string;
@@ -25,7 +25,15 @@ export class Form {
   form: HTMLElement;
   formControls: FormControl[];
 
-  /** Emitted when the form is submitted. */
+  /** Prevent the form from validating inputs before submitting. */
+  @Prop() novalidate = false;
+
+  /**
+   * Emitted when the form is submitted. This event will not be emitted if any form control inside of it is in an
+   * invalid state, unless the form has the `novalidate` attribute. Note that there is never a need to prevent this
+   * event, since it doen't send a GET or POST request like native forms. To "prevent" submission, use a conditional
+   * around the XHR request you use to submit the form's data with.
+   */
   @Event() slSubmit: EventEmitter;
 
   connectedCallback() {
@@ -188,13 +196,29 @@ export class Form {
       .filter(el => tags.includes(el.tagName.toLowerCase())) as HTMLElement[];
   }
 
-  /** Submits the form. */
+  /**
+   * Submits the form. If all controls are valid, the `slSubmit` event will be emitted and the promise will resolve with
+   * `true`. If any form control is invalid, the promise will resolve with `false` and no event will be emitted.
+   */
   @Method()
   async submit() {
     const formData = await this.getFormData();
     const formControls = await this.getFormControls();
+    const formControlsThatReport = formControls.filter((el: any) => typeof el.reportValidity === 'function') as any;
+
+    if (!this.novalidate) {
+      for (const el of formControlsThatReport) {
+        const isValid = await el.reportValidity();
+
+        if (!isValid) {
+          return false;
+        }
+      }
+    }
 
     this.slSubmit.emit({ formData, formControls });
+
+    return true;
   }
 
   handleClick(event: MouseEvent) {